Unify controllers and migrations of categories
