+++ routes/categories.js
@@ -1,13 +1,13 @@
 const express = require('express');
 const router = express.Router()
-const {getCategoryById, getCategories, updateCategories, createCategory, deleteCategory} = require('../controllers/categories.controller')
+const {getCategoryById, getCategories, updateCategories, getCategoriesName, deleteCategorie, createCategorie} = require('../controllers/categories.controller')
 
-router
-            .get('/:id', getCategoryById)
-            .get('/', getCategories)
-            .post('/', createCategory)
-            .put('/:categorieId', updateCategories)
-            .delete('/:id', deleteCategory)
+router.get('/name', getCategoriesName)
+router.get('/:id', getCategoryById)
+router.get('/', getCategories)
+router.post('/', createCategorie)
+router.put('/:categorieId', updateCategories)
+router.delete('/:id', deleteCategorie)
   
-  module.exports = router;
+module.exports = router;
